Implement component auto-configuration
