+++ tech.kage.event.kafka.reactor/src/test/java/tech/kage/event/kafka/reactor/MicrometerReactorKafkaEventStoreIT.java
@@ -45,10 +45,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.core.io.Resource;
 import org.springframework.kafka.config.TopicBuilder;
 import org.springframework.kafka.core.KafkaAdmin;
@@ -82,7 +81,9 @@ import tech.kage.event.Event;
  * 
  * @author Dariusz Szpakowski
  */
-@SpringBootTest
+@SpringBootTest(classes = {
+        ReactorKafkaEventStoreAutoConfiguration.class,
+        MicrometerReactorKafkaEventStoreIT.TestConfig.class })
 @ActiveProfiles("test")
 @Testcontainers
 @DirtiesContext
@@ -146,10 +147,9 @@ class MicrometerReactorKafkaEventStoreIT {
                 () -> "http://%s:%s".formatted(schemaRegistry.getHost(), schemaRegistry.getFirstMappedPort()));
     }
 
-    @Configuration
+    @TestConfiguration
     @EnableAutoConfiguration
-    @Import(ReactorKafkaEventStore.class)
-    static class TestConfiguration {
+    static class TestConfig {
         @Bean
         MeterRegistry meterRegistry() {
             return new SimpleMeterRegistry();
