+++ tech.kage.event.kafka.reactor/src/test/java/tech/kage/event/kafka/reactor/ReactorKafkaEventTransformerIT.java
@@ -70,9 +70,8 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Scope;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.convention.TestBean;
@@ -91,7 +90,10 @@ import tech.kage.event.crypto.EventEncryptor;
  * 
  * @author Dariusz Szpakowski
  */
-@SpringBootTest
+@SpringBootTest(classes = {
+        ReactorKafkaEventTransformer.class,
+        EventEncryptor.class,
+        ReactorKafkaEventTransformerIT.TestConfig.class })
 @ActiveProfiles("test")
 class ReactorKafkaEventTransformerIT {
     // UUT
@@ -134,9 +136,8 @@ class ReactorKafkaEventTransformerIT {
 
     static final Map<URI, KeysetHandle> testKms = new HashMap<>();
 
-    @Configuration
-    @Import({ ReactorKafkaEventTransformer.class, EventEncryptor.class })
-    static class TestConfiguration {
+    @TestConfiguration
+    static class TestConfig {
         @Bean
         @Scope(SCOPE_PROTOTYPE)
         Aead aead(URI encryptionKey) throws GeneralSecurityException {
