+++ tech.kage.event.kafka.streams/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+tech.kage.event.kafka.streams.KafkaStreamsEventStoreAutoConfiguration
