+++ tech.kage.event.postgres/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+tech.kage.event.postgres.PostgresEventStoreAutoConfiguration
