+++ tech.kage.event.postgres/src/test/java/tech/kage/event/postgres/ByteArrayKeyPostgresEventStoreIT.java
@@ -49,7 +49,7 @@ class ByteArrayKeyPostgresEventStoreIT extends PostgresEventStoreIT<byte[]> {
         return "bytea";
     }
 
-    static class TestConfiguration extends PostgresEventStoreIT.TestConfiguration {
+    static class TestConfig extends PostgresEventStoreIT.TestConfig {
     }
 
     static Stream<Arguments> testEvents() {
