+++ tech.kage.event.postgres/src/test/java/tech/kage/event/postgres/PostgresEventStoreIT.java
@@ -59,10 +59,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.core.io.Resource;
 import org.springframework.r2dbc.core.DatabaseClient;
 import org.springframework.test.context.ActiveProfiles;
@@ -135,10 +134,9 @@ abstract class PostgresEventStoreIT<K> {
                 () -> "http://%s:%s".formatted(schemaRegistry.getHost(), schemaRegistry.getFirstMappedPort()));
     }
 
-    @Configuration
+    @TestConfiguration
     @EnableAutoConfiguration
-    @Import(PostgresEventStore.class)
-    static class TestConfiguration {
+    static class TestConfig {
         private static final String DESERIALIZER_CLASS = "io.confluent.kafka.serializers.KafkaAvroDeserializer";
 
         @Bean
