+++ tech.kage.event.postgres/src/test/java/tech/kage/event/postgres/UUIDKeyPostgresEventStoreIT.java
@@ -49,7 +49,7 @@ class UUIDKeyPostgresEventStoreIT extends PostgresEventStoreIT<UUID> {
         return "uuid";
     }
 
-    static class TestConfiguration extends PostgresEventStoreIT.TestConfiguration {
+    static class TestConfig extends PostgresEventStoreIT.TestConfig {
     }
 
     static Stream<Arguments> testEvents() {
